Document raw string literals (#29079)
* Update existing samples

Make the edits for raw string literals to most content. For the programming guide content, move samples, format and update existing content.

* warnings, and move one more sample.

* first draft is completed.

* fix build warnings and proofread.

* fix snippet tags.

* Apply suggestions from code review


* respond to feedback.

* respond to feedback.

* move raw interpolated string literals to code

* Apply suggestions from code review


* respond to feedback.
