+++ docs/csharp/language-reference/tokens/snippets/string-interpolation.cs
@@ -101,4 +101,31 @@ public class StringInterpolation
         // </Newlines>
         Console.WriteLine(message);
     }
+
+    private static void InterpolatedRawLiteralStrings()
+    {
+        // <RawInterpolatedLiteralString>
+        int X = 2;
+        int Y = 3;
+
+        var pointMessage = $"""The point "{X}, {Y}" is {Math.Sqrt(X * X + Y * Y)} from the origin""";
+
+        Console.WriteLine(pointMessage);
+        // output:  The point "2, 3" is 3.605551275463989 from the origin.
+        // </RawInterpolatedLiteralString>
+
+    }
+    private static void InterpolatedRawLiteralStringsWithBraces()
+    {
+        // <RawInterpolatedLiteralStringWithBraces>
+        int X = 2;
+        int Y = 3;
+
+        var pointMessage = $$"""The point {{{X}}}, {{{Y}}} is {{Math.Sqrt(X * X + Y * Y)}} from the origin""";
+
+        Console.WriteLine(pointMessage);
+        // output:  The point {2, 3} is 3.605551275463989 from the origin.
+        // </RawInterpolatedLiteralStringWithBraces>
+
+    }
 }
