--- samples/snippets/csharp/VS_Snippets_VBCSharp/csProgGuideStrings/CS/CSRefStrings.csproj
@@ -1,16 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net48</TargetFramework>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <LangVersion>8.0</LangVersion>
-
-    <StartupObject>CsCsrefProgrammingStrings.ANSIToUnicode</StartupObject>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Reference Include="System.Windows.Forms" />
-  </ItemGroup>
-
-</Project>
