--- samples/snippets/csharp/VS_Snippets_VBCSharp/csProgGuideStrings/CS/TestStringBuilder.cs
@@ -1,26 +1,0 @@
-﻿using System;
-using System.Text;
-
-namespace CSRefStrings
-{
-    class TestStringBuilder
-    {
-        static void Main()
-        {
-            var sb = new StringBuilder();
-
-            // Create a string composed of numbers 0 - 9
-            for (int i = 0; i < 10; i++)
-            {
-                sb.Append(i.ToString());
-            }
-            Console.WriteLine(sb);  // displays 0123456789
-
-            // Copy one character of the string (not possible with a System.String)
-            sb[0] = sb[9];
-
-            Console.WriteLine(sb);  // displays 9123456789
-            Console.WriteLine();
-        }
-    }
-}
